Fixed some CDR mapping issues

+++ WWCP_OCPIv2.2.1/DataStructures/CDR.cs
@@ -373,7 +373,7 @@ namespace cloud.charging.open.protocols.OCPIv2_2_1
                            tariff.Currency,
                            chargingPeriods,
                            new Price(
-                               (Double) cdrCostDetails.TotalCost
+                               cdrCostDetails.TotalCost
                            ),
                            cdrCostDetails.TotalEnergy,
                            cdrCostDetails.TotalTime,
